expand options
make options for # vehicles, size of penalties, size of capacities,
etc, in order to reveal issues and corner cases.

with just two vehicles, it looked like the solver was correctly *not*
using A,B versions of vehicles.  But when you bumped up the number of
vehicles and bumped up the cost of using a combo, suddenly the best
solution used both single and a combo from one of the trucks, which is
forbidden, proving that you really do need to turn on that
--cumulative_constraint flag in order to trugger the combo_on *
single_on == 0 constraint.